type data 분리
