+++ src/components/Common/TextInput/SelectTypes.ts
@@ -1,0 +1,10 @@
+export interface Option {
+  value: string;
+  name: string;
+}
+
+export interface SelectBoxProps {
+  width: number | string;
+  options: Option[];
+  onChange: (value: string) => void;
+}
